Lots of fixes using Clang Static Analyzer

+++ Parser/parsingtreeconstiterator.cpp
@@ -14,12 +14,12 @@ ParsingTreeConstIterator::ParsingTreeConstIterator(const ParsingTree *tree) :
 
 ParsingTreeConstIterator &ParsingTreeConstIterator::goToChild(const unsigned int index)
 {
-    if(index >= (uint) currentNode->children.size())
+    if(index >= static_cast<uint>(currentNode->children.size()))
     {
         throw std::invalid_argument("This child does not exist.");
     }
 
-    currentNode = currentNode->children[index].get();
+    currentNode = currentNode->children[static_cast<int>(index)].get();
 
     return *this;
 }
@@ -112,7 +112,7 @@ QVector<unsigned int> ParsingTreeConstIterator::convertStringToPath(const QStrin
 
     std::for_each(coordinatesList.begin(), coordinatesList.end(), [&](const QString &str)
     {
-        pathVector.push_back(str.toInt());
+        pathVector.push_back(static_cast<uint>(str.toInt()));
     });
 
     return pathVector;
@@ -122,5 +122,5 @@ QString ParsingTreeConstIterator::removeOuterParenthesis(const QString &path) co
 {
     const unsigned int parenthesisPadding = 1;
     const unsigned int numberOfParenthesis = 2;
-    return path.mid(parenthesisPadding, path.size() - numberOfParenthesis * parenthesisPadding);
+    return path.mid(parenthesisPadding, static_cast<int>(static_cast<uint>(path.size()) - numberOfParenthesis * parenthesisPadding));
 }
